Code refactoring and seperating the implementation and function definition
