+++ src/main/java/com/example/springTutorial/dao/PersonsVehiclesDao.java
@@ -7,14 +7,9 @@ import java.util.Optional;
 
 public interface PersonsVehiclesDao {
     int insertPerson(PersonVehicles personVehicles);
-
     Optional<PersonVehicles> selectPersonByMobileNumber(String mobileNumber);
-
     int deletePersonById(String mobileNumber);
-
     Optional<QRCodeDetails> getQRScannedWeeklyQuota(String mobileNumber);
-
     int adjustWeeklyBalanceQuota(int balanceUnit, String id);
-
     int getWeeklyBalanceQuota(String id);
 }
